서버 측에서 연속적으로 오는 데이터를 처리 못 함

+++ Server/Noise.h
@@ -41,10 +41,10 @@ public :
     ~CNoise();
 
 public:
-	int GetValue(VECTOR vector);
+	float GetValue(VECTOR vector);
 
 private:
-    int Evaluate(VECTOR vector);
+    float Evaluate(VECTOR vector);
 
 private:
     void Randomize(int seed);
